Rename helper in demo
