+++ plugins/@grouparoo/demo/__tests__/bin/reset.ts
@@ -1,11 +1,4 @@
-import path from "path";
-process.env.GROUPAROO_INJECTED_PLUGINS = JSON.stringify({
-  "@grouparoo/postgres": {
-    path: path.join(__dirname, "..", "..", "..", "postgres"),
-  },
-});
-
-import { hooks } from "../utils/shared";
+import { hooks } from "../utils/helper";
 import { Demo } from "../../src/bin/grouparoo/demo/demo";
 import {
   Source,
